Extract Modal, Row, Column components to own files
